+++ CHANGELOG.md
@@ -11,12 +11,17 @@ This release introduces the following breaking changes:
 
 - feat(sumologicextension): use hostname as default collector name [#918]
 
+### Added
+
+- feat(sourceprocessor): add debug logs for source category filler [#944]
+
 ### Fixed
 
 - fix(k8sprocessor): race condition when getting Pod data [#938]
 
 [#918]: https://github.com/SumoLogic/sumologic-otel-collector/pull/918
 [#938]: https://github.com/SumoLogic/sumologic-otel-collector/pull/938
+[#944]: https://github.com/SumoLogic/sumologic-otel-collector/pull/944
 [Unreleased]: https://github.com/SumoLogic/sumologic-otel-collector/compare/v0.70.0-sumo-0...main
 
 ## [v0.70.0-sumo-1]